Refactor src/ directory (#436)
* Move some libraries to src/utils

Verified that demo.conf still works.

* Move retokenize library

* Fix imports

* Move serialize.py

* Move config.py

* Move tasks.py

* Move most modules to modules/

* Move edge probing and seq2seq decoder modules
