+++ src/preprocess.py
@@ -31,11 +31,11 @@ except BaseException:
 
 import _pickle as pkl  # :(
 
-from . import config
-from . import serialize
-from . import utils
-from . import tasks as tasks_module
-from .tasks import MTTask
+from .utils import config
+from .utils import serialize
+from .utils import utils
+from .tasks import tasks as tasks_module
+from .tasks.tasks import MTTask
 
 ALL_GLUE_TASKS = ['sst', 'cola', 'mrpc', 'qqp', 'sts-b',
                   'mnli', 'qnli', 'rte', 'wnli', 'mnli-diagnostic']
